+++ CHANGELOG.md
@@ -13,6 +13,10 @@ Notable changes to Sparkle will be documented in this file.
 ### Added
 - Added documentation on how to use Venv instead of Conda for Sparkle [SPRK-367]
 - Added autocompletion for Sparkle CLI [SPRK-332]
+- Cancelling a job through the interactive table now has a confirmation dialog [SPRK-343]
+
+### Changed
+- The wait command and cancel command have been merged into the `jobs` command [SPRK-343]
 
 ## [0.9.1.2] - 2025/12/12